Moved all graphql from articles lib to separate file
